20170526_HOME_V1

+++ 20170525_29.php
@@ -1,4 +1,6 @@
 <?php
+//後端畫布(rate時記得去URL後面輸入?RATE=50)
+
 if(!isset($_GET['rate'])) exit(0);
 
 $rate =$_GET['rate'];
@@ -8,15 +10,14 @@ $rate =$_GET['rate'];
 $image = imagecreate("400","40");
 
 
-$yellow=imagecolorallocate ( $image , 255, 255 , 0);
+$yellow =imagecolorallocate ( $image , 255, 255 , 0);
 
-$red=imagecolorallocate ( $image , 255, 0, 0);
+$red = imagecolorallocate ( $image , 255, 0, 0);
 
-//int imageloadfont ( string $file )
+
 imagefilledrectangle ( $image , 0 , 0 , 400 , 40 , $yellow );
-imagefilledrectangle ( $image , 0 , 0 , (int)($rate*400/100) , 40 , $red );
+imagefilledrectangle ( $image , 0 , 0 , $rate*400/100 , 40 , $red );
 
-//imagestring ( $image , int $font , int $x , int $y , string $string , int $color )
 
 header('Content-Type:image/jpeg');
